requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-swagger==0.2.2
drf-yasg==1.21.5
Pillow==9.4.0
varname==0.11.0
